+++ pkg/controllers/v1alpha1/fluidapp/implement_test.go
@@ -20,17 +20,18 @@ import (
 	"testing"
 
 	"github.com/brahma-adshonor/gohook"
-	"github.com/fluid-cloudnative/fluid/pkg/common"
-	"github.com/fluid-cloudnative/fluid/pkg/utils/fake"
-	"github.com/fluid-cloudnative/fluid/pkg/utils/kubeclient"
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/fluid-cloudnative/fluid/pkg/common"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/fake"
+	"github.com/fluid-cloudnative/fluid/pkg/utils/kubeclient"
 )
 
-func TestFluidAppReconcilerImplement_umountFuseSidecar(t *testing.T) {
+func TestFluidAppReconcilerImplement_umountFuseSidecars(t *testing.T) {
 	mockExec := func(p1, p2, p3 string, p4 []string) (stdout string, stderr string, e error) {
 		return "", "", nil
 	}
@@ -78,7 +79,7 @@ func TestFluidAppReconcilerImplement_umountFuseSidecar(t *testing.T) {
 				pod: &corev1.Pod{
 					ObjectMeta: metav1.ObjectMeta{Name: "test"},
 					Spec: corev1.PodSpec{
-						Containers: []corev1.Container{{Name: common.FuseContainerName}},
+						Containers: []corev1.Container{{Name: common.FuseContainerName + "-0"}},
 					},
 				},
 			},
@@ -91,7 +92,7 @@ func TestFluidAppReconcilerImplement_umountFuseSidecar(t *testing.T) {
 					ObjectMeta: metav1.ObjectMeta{Name: "test"},
 					Spec: corev1.PodSpec{
 						Containers: []corev1.Container{{
-							Name: common.FuseContainerName,
+							Name: common.FuseContainerName + "-0",
 							Lifecycle: &corev1.Lifecycle{
 								PreStop: &corev1.LifecycleHandler{
 									Exec: &corev1.ExecAction{Command: []string{"umount"}},
@@ -110,12 +111,39 @@ func TestFluidAppReconcilerImplement_umountFuseSidecar(t *testing.T) {
 					ObjectMeta: metav1.ObjectMeta{Name: "test"},
 					Spec: corev1.PodSpec{
 						Containers: []corev1.Container{{
-							Name: common.FuseContainerName,
+							Name: common.FuseContainerName + "-0",
 							VolumeMounts: []corev1.VolumeMount{{
 								Name:      "juicefs-fuse-mount",
 								MountPath: "/mnt/jfs",
 							}},
 						}},
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "test-multi-sidecar",
+			args: args{
+				pod: &corev1.Pod{
+					ObjectMeta: metav1.ObjectMeta{Name: "test"},
+					Spec: corev1.PodSpec{
+						Containers: []corev1.Container{
+							{
+								Name: common.FuseContainerName + "-0",
+								VolumeMounts: []corev1.VolumeMount{{
+									Name:      "juicefs-fuse-mount",
+									MountPath: "/mnt/jfs",
+								}},
+							},
+							{
+								Name: common.FuseContainerName + "-1",
+								VolumeMounts: []corev1.VolumeMount{{
+									Name:      "juicefs-fuse-mount",
+									MountPath: "/mnt/jfs",
+								}},
+							},
+						},
 					},
 				},
 			},
@@ -127,7 +155,7 @@ func TestFluidAppReconcilerImplement_umountFuseSidecar(t *testing.T) {
 			i := &FluidAppReconcilerImplement{
 				Log: fake.NullLogger(),
 			}
-			if err := i.umountFuseSidecar(tt.args.pod); (err != nil) != tt.wantErr {
+			if err := i.umountFuseSidecars(tt.args.pod); (err != nil) != tt.wantErr {
 				t.Errorf("umountFuseSidecar() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})